Made a request for the year, month and variables of a remote slightly more efficient.
Avoided making separate calls for the request dict of the remote to get
the year and the month. Apparently, the request dict is pulled from the
server every time you access the Remote.request property. With the
current code, calling `remotes_dict_by_yearmonth` will make two such
requests (one for the variables and one for the year+month). To cut this
down to just one, we would have to not go through the separate
`get_remote_varset` and `get_remote_yearmonth` functions, and instead
just access `Remote.request` once for each Remote instance.